backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.0.1
coverage==7.2.1
cryptography==39.0.2
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.53.1
django-cors-headers==3.14.0
django-nose==1.4.7
django-phonenumber-field==7.0.2
django-random-id-model==0.1.1
django-rename-app==0.1.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
greenlet==1.1.3.post0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
msgpack==1.0.4
nose==1.3.7
oauthlib==3.2.2
phonenumbers==8.13.8
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pynvim==0.4.3
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
